merge dev

+++ modules/sdnq/layers/linear/linear_fp8_tensorwise.py
@@ -31,7 +31,7 @@ def fp8_matmul_tensorwise(
     if svd_up is not None:
         input.flatten(0,-2)
         if bias is not None:
-            bias = torch.addmm(bias, torch.mm(input.to(dtype=svd_down.dtype), svd_down), svd_up)
+            bias = torch.addmm(bias.to(dtype=svd_down.dtype), torch.mm(input.to(dtype=svd_down.dtype), svd_down), svd_up)
         else:
             bias = torch.mm(torch.mm(input.to(dtype=svd_down.dtype), svd_down), svd_up)
     dummy_input_scale = torch.ones(1, device=input.device, dtype=torch.float32)
